+++ index.js
@@ -82,7 +82,7 @@ function parse(o){
   }
   if (o.options.killBreaks) {
     // replace <br />(blanks goes here) to <br />.
-    o.html = o.html.replace(/(<br\s*\/?>(\s|&nbsp;?)*){1,}/g, '<br />');
+    o.html = o.html.replace(/<br[^\/>]*\/?>/ig, '<br />');
     // remove tab symbols like \r\t\n
     o.html = o.html.replace(/[\n\r\t]{2,}/gi, ' ');
   }